feat!: Cargo.toml updating
The "cargo_lock" updater has been renamed to "cargo". It will now update
a package's `Cargo.toml` in addition to `Cargo.lock`. This takes
advantage of the use of cargo metadata in order to find all of the
packages that belong in the workspace.

In order to support this, the `Replacer` trait has been modified to
return a Vec of replacers instead of a single item. This means that each
replacer can have multiple files that it updates.