test: Added test for cargo workspaces
This introduces a test that will check *bomper* against a workspace
consisting of three crates. Two of the crates use an inherited version
while the third does not.

While implementing this the some of the replacer logic was updated to
ensure that the correct files were updated.

The test is not thorough and does not check the contents, but only that
the correct files were generated. The rest will need to come later.