feat(cli): search for configuration file in project base directory
The previous behavior was to look for "bomp.ron" in the current
directory. With this change the program will now try to search for the
root directory of the current project for the configuration file. This
uses the `project-base-directory` crate which will check for the
`PRJ_ROOT` environment variable and use that path if set. If it is not
set then it will attempt to find the root directory by checking for the
git root. Failing that, it will fall back to the existing behavior.