feat: add `flakeModules.bomper` output
This flake-parts module can be used to create a wrapped version of
bomper with configuration pre-specified. This allows for usage of bomper
without commiting a configuration file into the repository. Simply
specify the configuration as a string and the module will created a
wrapped version of bomper that passes the configuration to it as its
first argument and passes the rest through.

This has removed the need for the flake to depend on itself by
dogfooding the module. See `flake-parts/bomper.nix` for an example of
how to use this.