feat: construct changelog entry from commits
This is a minimal example of building a changelog that is roughly
equivalent to the one that exists in this project. Since the changelog
benefits from having some extra information available about the commit,
e.g. the author details and commit hash, the functions have been changed
to take a wrapped `Commit` struct that stores the extra information as
well as the original `ConventionalCommit` struct.

This changelog is not quite as pretty as the existing one, doesn't
handle creating links for remote repositories, and doesn't link to
author profiles. This is seen as a nice to have but outside the scope of
this commit.