fix: attempt to always run from base directory of project
This should make sure that it isn't possible to try to perform file
replacement when located in the wrong directory.

Also contains a small tweak to configuration loading. It will now
attempt to look in the PRJ_CONFIG_HOME directory for a configuration
file first, only falling back to the current directory (which is now the
project root) if it fails to find one.