feat(cli/changelog): add `at` flag
This flag can be used to manually specify a version for which you wish
to display a changelog for.

This is required because the existing `--only-current-version` flag is
insufficient for displaying the changelog for a project that has already
bumped its version. In that case, it would be unable to print the
changes. I use some automation which needs this kind of functionality in
order to send version-specific changelog entries to the GitHub releases
page.