refactor(cli/changelog): use `unreleased` when showing un-comitted changes
This command originally required the user to specify what kind of
version increment they wished to use, using the same logic from the
`bump` command. However this is feels overkill for just seeing the
unrealeased changes.

This changes the changelog generation function to accept version as an
`&str`, which is fine because it was never doing anything that required
it to be a `semver::Version`.