feat: support repos with no existing tags
Repos with no existing tags will instead use the root commit as the
start, and behave as if it was tagged with `0.0.0`. This allows for a
changelog to be generated from everything but the first commit. For my
projects, I do an empty commit as the first one so this is perfect.