fix: commits when multiple files share the same name
There was an issue setting up the git commit when multiple `Cargo.toml`
files were present in a repository. This is because the logic that
created the commit would replace all of them. This fix works by building
a path recursively when navigating the tree.

This absolutely needs to be removed when gitoxide supports committing
directly.